% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanotime.R
\name{all.equal.nanotime}
\alias{all.equal.nanotime}
\alias{all.equal,nanotime,nanotime-method}
\alias{all.equal,nanotime,ANY-method}
\alias{all.equal,ANY,nanotime-method}
\title{Test if Two Objects are (Nearly) Equal}
\usage{
\method{all.equal}{nanotime}(target, current, ...)

\S4method{all.equal}{nanotime,nanotime}(target, current, ...)

\S4method{all.equal}{nanotime,ANY}(target, current, ...)

\S4method{all.equal}{ANY,nanotime}(target, current, ...)
}
\arguments{
\item{target, current}{\code{nanotime} arguments to be compared}

\item{...}{further arguments for different methods}
}
\description{
Compare \code{target} and \code{current} testing \sQuote{near
equality}.  If they are different, comparison is still made to
some extent, and a report of the differences is returned.  Do not
use \code{all.equal} directly in \code{if} expressions---either
use \code{isTRUE(all.equal(....))} or \code{\link{identical}} if
appropriate.
}
\seealso{
\code{\link{identical}}, \code{\link{isTRUE}},
    \code{\link{==}}, and \code{\link{all}} for exact equality
    testing.
}
