% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration.R
\docType{class}
\name{duration-class}
\alias{duration-class}
\alias{duration}
\alias{as.duration,character-method}
\alias{as.duration,integer64-method}
\alias{as.duration,numeric-method}
\alias{as.duration,integer-method}
\alias{show,duration-method}
\alias{print,duration-method}
\alias{format.duration}
\alias{-,duration,duration-method}
\alias{-,duration,integer64-method}
\alias{-,duration,integer-method}
\alias{-,duration,numeric-method}
\alias{-,duration,ANY-method}
\alias{-,integer64,duration-method}
\alias{-,integer,duration-method}
\alias{-,numeric,duration-method}
\alias{-,ANY,duration-method}
\alias{+,duration,ANY-method}
\alias{+,duration,duration-method}
\alias{+,duration,integer64-method}
\alias{+,duration,numeric-method}
\alias{+,ANY,duration-method}
\alias{+,integer64,duration-method}
\alias{+,numeric,duration-method}
\alias{*,duration,duration-method}
\alias{*,duration,numeric-method}
\alias{*,duration,integer64-method}
\alias{*,numeric,duration-method}
\alias{*,integer64,duration-method}
\alias{*,duration,ANY-method}
\alias{*,ANY,duration-method}
\alias{/,duration,duration-method}
\alias{/,duration,integer64-method}
\alias{/,duration,numeric-method}
\alias{/,duration,ANY-method}
\alias{/,ANY,duration-method}
\alias{Arith,duration,ANY-method}
\alias{Compare,duration,ANY-method}
\alias{Logic,duration,duration-method}
\alias{Logic,duration,ANY-method}
\alias{Logic,ANY,duration-method}
\alias{abs,duration-method}
\alias{sign,duration-method}
\alias{Math,duration-method}
\alias{Math2,duration-method}
\alias{Summary,duration-method}
\alias{sum,duration-method}
\alias{min,duration-method}
\alias{max,duration-method}
\alias{range,duration-method}
\alias{Complex,duration-method}
\alias{[[,duration-method}
\alias{[,duration,ANY-method}
\alias{[<-,duration,ANY,ANY,ANY-method}
\alias{c.duration}
\alias{names<-,duration-method}
\title{Duration type with nanosecond precision}
\usage{
duration(hours, minutes, seconds, nanoseconds)

\S4method{as.duration}{character}(x)

\S4method{as.duration}{integer64}(x)

\S4method{as.duration}{numeric}(x)

\S4method{as.duration}{integer}(x)

\S4method{show}{duration}(object)

\S4method{print}{duration}(x, ...)

\method{format}{duration}(x, ...)

\S4method{-}{duration,duration}(e1, e2)

\S4method{-}{duration,integer64}(e1, e2)

\S4method{-}{duration,integer}(e1, e2)

\S4method{-}{duration,numeric}(e1, e2)

\S4method{-}{duration,ANY}(e1, e2)

\S4method{-}{integer64,duration}(e1, e2)

\S4method{-}{integer,duration}(e1, e2)

\S4method{-}{numeric,duration}(e1, e2)

\S4method{-}{ANY,duration}(e1, e2)

\S4method{+}{duration,ANY}(e1, e2)

\S4method{+}{duration,duration}(e1, e2)

\S4method{+}{duration,integer64}(e1, e2)

\S4method{+}{duration,numeric}(e1, e2)

\S4method{+}{ANY,duration}(e1, e2)

\S4method{+}{integer64,duration}(e1, e2)

\S4method{+}{numeric,duration}(e1, e2)

\S4method{*}{duration,duration}(e1, e2)

\S4method{*}{duration,numeric}(e1, e2)

\S4method{*}{duration,integer64}(e1, e2)

\S4method{*}{numeric,duration}(e1, e2)

\S4method{*}{integer64,duration}(e1, e2)

\S4method{*}{duration,ANY}(e1, e2)

\S4method{*}{ANY,duration}(e1, e2)

\S4method{/}{duration,duration}(e1, e2)

\S4method{/}{duration,integer64}(e1, e2)

\S4method{/}{duration,numeric}(e1, e2)

\S4method{/}{duration,ANY}(e1, e2)

\S4method{/}{ANY,duration}(e1, e2)

\S4method{Arith}{duration,ANY}(e1, e2)

\S4method{Compare}{duration,ANY}(e1, e2)

\S4method{Logic}{duration,duration}(e1, e2)

\S4method{Logic}{duration,ANY}(e1, e2)

\S4method{Logic}{ANY,duration}(e1, e2)

\S4method{abs}{duration}(x)

\S4method{sign}{duration}(x)

\S4method{Math}{duration}(x)

\S4method{Math2}{duration}(x, digits)

\S4method{Summary}{duration}(x, ..., na.rm = FALSE)

\S4method{sum}{duration}(x, ..., na.rm = FALSE)

\S4method{min}{duration}(x, ..., na.rm = FALSE)

\S4method{max}{duration}(x, ..., na.rm = FALSE)

\S4method{range}{duration}(x, ..., na.rm = FALSE)

\S4method{Complex}{duration}(z)

\S4method{[[}{duration}(x, i, j, ..., drop = FALSE)

\S4method{[}{duration,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{duration,ANY,ANY,ANY}(x, i, j, ...) <- value

\method{c}{duration}(...)

\S4method{names}{duration}(x) <- value
}
\arguments{
\item{hours}{number of hours}

\item{minutes}{number of minutes}

\item{seconds}{number of seconds}

\item{nanoseconds}{number of nanoseconds}

\item{x}{a \code{duration} object}

\item{object}{argument for method \code{show}}

\item{...}{further arguments passed to or from methods.}

\item{e1}{Operand of class \code{nanoival}}

\item{e2}{Operand of class \code{nanoival}}

\item{digits}{Required for \code{Math2} signature but ignored here}

\item{z}{Required for \code{Complex} signature but ignored here}

\item{i}{index specifying elements to extract or replace.}

\item{j}{Required for \code{[} signature but ignored here}

\item{drop}{Required for \code{[} signature but ignored here}

\item{value}{argument for \code{duration-class}}

\item{tz}{a timezone string}

\item{recursive}{argument for method \code{c}}
}
\value{
A duration object
}
\description{
\code{duration} is a length of time type (implemented as an S4
class) with nanosecond precision. It is a count of nanoseconds
which may be negative.
}
\details{
A duration can be constructed with the function 'as.duration'
which can take the types 'integer64', 'integer' and 'numeric' (all
indicating the count in nanosecond units) or the type 'character'.
}
\section{Output Format}{


A \code{duration} is displayed as hours, minutes, seconds and
nanoseconds like this: \code{110:12:34.123_453_000}.
}

\author{
Leonardo Silvestri
}
