% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoival.R
\name{is.unsorted,nanoival-method}
\alias{is.unsorted,nanoival-method}
\title{Test if a \code{nanoival} vector is Not Sorted}
\usage{
\S4method{is.unsorted}{nanoival}(x, na.rm = FALSE, strictly = FALSE)
}
\arguments{
\item{x}{a \code{nanoival} vector}

\item{na.rm}{logical.  Should missing values be removed before
checking?}

\item{strictly}{logical indicating if the check should be for
_strictly_ increasing values.}
}
\description{
Test if an object is not sorted (in increasing order), without the
cost of sorting it.
}
\seealso{
\code{\link{sort}}
}
