% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoduration.R
\docType{class}
\name{nanoduration-class}
\alias{nanoduration-class}
\alias{nanoduration}
\alias{as.nanoduration,character-method}
\alias{as.nanoduration,integer64-method}
\alias{as.nanoduration,numeric-method}
\alias{as.nanoduration,integer-method}
\alias{show,nanoduration-method}
\alias{print,nanoduration-method}
\alias{format.nanoduration}
\alias{as.character,nanoduration-method}
\alias{is.na,nanoduration-method}
\alias{-,nanoduration,nanoduration-method}
\alias{-,nanoduration,integer64-method}
\alias{-,nanoduration,integer-method}
\alias{-,nanoduration,numeric-method}
\alias{-,nanoduration,ANY-method}
\alias{-,nanotime,nanoduration-method}
\alias{-,integer64,nanoduration-method}
\alias{-,integer,nanoduration-method}
\alias{-,numeric,nanoduration-method}
\alias{-,ANY,nanoduration-method}
\alias{+,nanoduration,ANY-method}
\alias{+,nanoduration,nanoduration-method}
\alias{+,nanoduration,integer64-method}
\alias{+,nanoduration,numeric-method}
\alias{+,nanotime,nanoduration-method}
\alias{+,ANY,nanoduration-method}
\alias{+,integer64,nanoduration-method}
\alias{+,numeric,nanoduration-method}
\alias{*,nanoduration,nanoduration-method}
\alias{*,nanoduration,numeric-method}
\alias{*,nanoduration,integer64-method}
\alias{*,numeric,nanoduration-method}
\alias{*,integer64,nanoduration-method}
\alias{*,nanoduration,ANY-method}
\alias{*,ANY,nanoduration-method}
\alias{/,nanoduration,nanoduration-method}
\alias{/,nanoduration,integer64-method}
\alias{/,nanoduration,numeric-method}
\alias{/,nanoduration,ANY-method}
\alias{/,ANY,nanoduration-method}
\alias{Arith,nanoduration,ANY-method}
\alias{Compare,nanoduration,ANY-method}
\alias{Logic,nanoduration,nanoduration-method}
\alias{Logic,nanoduration,ANY-method}
\alias{Logic,ANY,nanoduration-method}
\alias{abs,nanoduration-method}
\alias{sign,nanoduration-method}
\alias{Math,nanoduration-method}
\alias{Math2,nanoduration-method}
\alias{Summary,nanoduration-method}
\alias{sum,nanoduration-method}
\alias{min,nanoduration-method}
\alias{max,nanoduration-method}
\alias{range,nanoduration-method}
\alias{Complex,nanoduration-method}
\alias{[[,nanoduration-method}
\alias{[,nanoduration,ANY-method}
\alias{[<-,nanoduration,ANY,ANY,ANY-method}
\alias{c.nanoduration}
\alias{names<-,nanoduration-method}
\title{Duration type with nanosecond precision}
\usage{
nanoduration(hours, minutes, seconds, nanoseconds)

\S4method{as.nanoduration}{character}(x)

\S4method{as.nanoduration}{integer64}(x)

\S4method{as.nanoduration}{numeric}(x)

\S4method{as.nanoduration}{integer}(x)

\S4method{show}{nanoduration}(object)

\S4method{print}{nanoduration}(x, ...)

\method{format}{nanoduration}(x, ...)

\S4method{as.character}{nanoduration}(x)

\S4method{is.na}{nanoduration}(x)

\S4method{-}{nanoduration,nanoduration}(e1, e2)

\S4method{-}{nanoduration,integer64}(e1, e2)

\S4method{-}{nanoduration,integer}(e1, e2)

\S4method{-}{nanoduration,numeric}(e1, e2)

\S4method{-}{nanoduration,ANY}(e1, e2)

\S4method{-}{nanotime,nanoduration}(e1, e2)

\S4method{-}{integer64,nanoduration}(e1, e2)

\S4method{-}{integer,nanoduration}(e1, e2)

\S4method{-}{numeric,nanoduration}(e1, e2)

\S4method{-}{ANY,nanoduration}(e1, e2)

\S4method{+}{nanoduration,ANY}(e1, e2)

\S4method{+}{nanoduration,nanoduration}(e1, e2)

\S4method{+}{nanoduration,integer64}(e1, e2)

\S4method{+}{nanoduration,numeric}(e1, e2)

\S4method{+}{nanotime,nanoduration}(e1, e2)

\S4method{+}{ANY,nanoduration}(e1, e2)

\S4method{+}{integer64,nanoduration}(e1, e2)

\S4method{+}{numeric,nanoduration}(e1, e2)

\S4method{*}{nanoduration,nanoduration}(e1, e2)

\S4method{*}{nanoduration,numeric}(e1, e2)

\S4method{*}{nanoduration,integer64}(e1, e2)

\S4method{*}{numeric,nanoduration}(e1, e2)

\S4method{*}{integer64,nanoduration}(e1, e2)

\S4method{*}{nanoduration,ANY}(e1, e2)

\S4method{*}{ANY,nanoduration}(e1, e2)

\S4method{/}{nanoduration,nanoduration}(e1, e2)

\S4method{/}{nanoduration,integer64}(e1, e2)

\S4method{/}{nanoduration,numeric}(e1, e2)

\S4method{/}{nanoduration,ANY}(e1, e2)

\S4method{/}{ANY,nanoduration}(e1, e2)

\S4method{Arith}{nanoduration,ANY}(e1, e2)

\S4method{Compare}{nanoduration,ANY}(e1, e2)

\S4method{Logic}{nanoduration,nanoduration}(e1, e2)

\S4method{Logic}{nanoduration,ANY}(e1, e2)

\S4method{Logic}{ANY,nanoduration}(e1, e2)

\S4method{abs}{nanoduration}(x)

\S4method{sign}{nanoduration}(x)

\S4method{Math}{nanoduration}(x)

\S4method{Math2}{nanoduration}(x, digits)

\S4method{Summary}{nanoduration}(x, ..., na.rm = FALSE)

\S4method{sum}{nanoduration}(x, ..., na.rm = FALSE)

\S4method{min}{nanoduration}(x, ..., na.rm = FALSE)

\S4method{max}{nanoduration}(x, ..., na.rm = FALSE)

\S4method{range}{nanoduration}(x, ..., na.rm = FALSE)

\S4method{Complex}{nanoduration}(z)

\S4method{[[}{nanoduration}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoduration,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoduration,ANY,ANY,ANY}(x, i, j, ...) <- value

\method{c}{nanoduration}(...)

\S4method{names}{nanoduration}(x) <- value
}
\arguments{
\item{hours}{number of hours}

\item{minutes}{number of minutes}

\item{seconds}{number of seconds}

\item{nanoseconds}{number of nanoseconds}

\item{x}{a \code{nanoduration} object}

\item{object}{argument for method \code{show}}

\item{...}{further arguments passed to or from methods.}

\item{e1}{Operand of class \code{nanoival}}

\item{e2}{Operand of class \code{nanoival}}

\item{digits}{Required for \code{Math2} signature but ignored here}

\item{z}{Required for \code{Complex} signature but ignored here}

\item{i}{index specifying elements to extract or replace.}

\item{j}{Required for \code{[} signature but ignored here}

\item{drop}{Required for \code{[} signature but ignored here}

\item{value}{argument for \code{nanoduration-class}}

\item{tz}{a timezone string}

\item{recursive}{argument for method \code{c}}
}
\value{
A nanoduration object
}
\description{
\code{nanoduration} is a length of time type (implemented as an S4
class) with nanosecond precision. It is a count of nanoseconds
which may be negative.
}
\details{
A nanoduration can be constructed with the function 'as.nanoduration'
which can take the types 'integer64', 'integer' and 'numeric' (all
indicating the count in nanosecond units) or the type 'character'.
}
\section{Output Format}{


A \code{nanoduration} is displayed as hours, minutes, seconds and
nanoseconds like this: \code{110:12:34.123_453_000}.
}

\author{
Leonardo Silvestri
}
