% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoperiod.R
\docType{class}
\name{nanoperiod-class}
\alias{nanoperiod-class}
\alias{nanoperiod}
\alias{as.nanoperiod,character-method}
\alias{as.nanoperiod,integer64-method}
\alias{as.nanoperiod,numeric-method}
\alias{as.nanoperiod,integer-method}
\alias{as.nanoperiod,nanoduration-method}
\alias{show,nanoperiod-method}
\alias{print,nanoperiod-method}
\alias{format.nanoperiod}
\alias{as.character,nanoperiod-method}
\alias{is.na,nanoperiod-method}
\alias{is.na<-,nanoperiod-method}
\alias{[[,nanoperiod-method}
\alias{[,nanoperiod,ANY-method}
\alias{[<-,nanoperiod,ANY,ANY,ANY-method}
\alias{c.nanoperiod}
\alias{names,nanoperiod-method}
\alias{names<-,nanoperiod-method}
\alias{Ops,nanoperiod,ANY-method}
\alias{Ops,ANY,nanoperiod-method}
\alias{Math,nanoperiod-method}
\alias{Math2,nanoperiod-method}
\alias{Summary,nanoperiod-method}
\alias{Complex,nanoperiod-method}
\alias{-,nanoperiod,ANY-method}
\alias{-,nanoperiod,nanoperiod-method}
\alias{-,nanoperiod,nanoduration-method}
\alias{-,nanoperiod,integer64-method}
\alias{-,nanotime,nanoperiod-method}
\alias{+,nanoperiod,nanotime-method}
\alias{-,nanoperiod,numeric-method}
\alias{-,ANY,nanoperiod-method}
\alias{-,nanoduration,nanoperiod-method}
\alias{-,integer64,nanoperiod-method}
\alias{-,numeric,nanoperiod-method}
\alias{+,nanoperiod,ANY-method}
\alias{+,nanoperiod,nanoperiod-method}
\alias{+,ANY,nanoperiod-method}
\alias{+,nanoperiod,nanoduration-method}
\alias{+,nanoperiod,integer64-method}
\alias{+,nanotime,nanoperiod-method}
\alias{+,nanoperiod,numeric-method}
\alias{+,nanoduration,nanoperiod-method}
\alias{+,integer64,nanoperiod-method}
\alias{+,numeric,nanoperiod-method}
\alias{*,ANY,nanoperiod-method}
\alias{*,nanoperiod,ANY-method}
\alias{*,nanoperiod,integer64-method}
\alias{*,nanoperiod,numeric-method}
\alias{*,integer64,nanoperiod-method}
\alias{*,numeric,nanoperiod-method}
\alias{/,ANY,nanoperiod-method}
\alias{/,nanoperiod,ANY-method}
\alias{/,nanoperiod,integer64-method}
\alias{/,nanoperiod,numeric-method}
\alias{Compare,ANY,nanoperiod-method}
\alias{Compare,nanoperiod,ANY-method}
\alias{==,nanoperiod,nanoperiod-method}
\alias{!=,nanoperiod,nanoperiod-method}
\title{Period type with nanosecond precision}
\usage{
nanoperiod(months, days, duration = as.nanoduration(0))

\S4method{as.nanoperiod}{character}(x)

\S4method{as.nanoperiod}{integer64}(x)

\S4method{as.nanoperiod}{numeric}(x)

\S4method{as.nanoperiod}{integer}(x)

\S4method{as.nanoperiod}{nanoduration}(x)

\S4method{show}{nanoperiod}(object)

\S4method{print}{nanoperiod}(x, ...)

\method{format}{nanoperiod}(x, ...)

\S4method{as.character}{nanoperiod}(x)

\S4method{is.na}{nanoperiod}(x)

\S4method{is.na}{nanoperiod}(x) <- value

\S4method{[[}{nanoperiod}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoperiod,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{nanoperiod,ANY,ANY,ANY}(x, i, j, ...) <- value

\method{c}{nanoperiod}(...)

\S4method{names}{nanoperiod}(x)

\S4method{names}{nanoperiod}(x) <- value

\S4method{Ops}{nanoperiod,ANY}(e1, e2)

\S4method{Ops}{ANY,nanoperiod}(e1, e2)

\S4method{Math}{nanoperiod}(x)

\S4method{Math2}{nanoperiod}(x, digits)

\S4method{Summary}{nanoperiod}(x, ..., na.rm = FALSE)

\S4method{Complex}{nanoperiod}(z)

\S4method{-}{nanoperiod,ANY}(e1, e2)

\S4method{-}{nanoperiod,nanoperiod}(e1, e2)

\S4method{-}{nanoperiod,ANY}(e1, e2)

\S4method{-}{nanoperiod,nanoduration}(e1, e2)

\S4method{-}{nanoperiod,integer64}(e1, e2)

\S4method{-}{nanotime,nanoperiod}(e1, e2)

\S4method{+}{nanoperiod,nanotime}(e1, e2)

\S4method{-}{nanoperiod,numeric}(e1, e2)

\S4method{-}{ANY,nanoperiod}(e1, e2)

\S4method{-}{nanoduration,nanoperiod}(e1, e2)

\S4method{-}{integer64,nanoperiod}(e1, e2)

\S4method{-}{numeric,nanoperiod}(e1, e2)

\S4method{+}{nanoperiod,ANY}(e1, e2)

\S4method{+}{nanoperiod,nanoperiod}(e1, e2)

\S4method{+}{ANY,nanoperiod}(e1, e2)

\S4method{+}{nanoperiod,nanoduration}(e1, e2)

\S4method{+}{nanoperiod,integer64}(e1, e2)

\S4method{+}{nanotime,nanoperiod}(e1, e2)

\S4method{+}{nanoperiod,numeric}(e1, e2)

\S4method{+}{nanoperiod,ANY}(e1, e2)

\S4method{+}{nanoduration,nanoperiod}(e1, e2)

\S4method{+}{integer64,nanoperiod}(e1, e2)

\S4method{+}{numeric,nanoperiod}(e1, e2)

\S4method{*}{ANY,nanoperiod}(e1, e2)

\S4method{*}{nanoperiod,ANY}(e1, e2)

\S4method{*}{nanoperiod,integer64}(e1, e2)

\S4method{*}{nanoperiod,numeric}(e1, e2)

\S4method{*}{integer64,nanoperiod}(e1, e2)

\S4method{*}{numeric,nanoperiod}(e1, e2)

\S4method{/}{ANY,nanoperiod}(e1, e2)

\S4method{/}{nanoperiod,ANY}(e1, e2)

\S4method{/}{nanoperiod,integer64}(e1, e2)

\S4method{/}{nanoperiod,numeric}(e1, e2)

\S4method{Compare}{ANY,nanoperiod}(e1, e2)

\S4method{Compare}{nanoperiod,ANY}(e1, e2)

\S4method{==}{nanoperiod,nanoperiod}(e1, e2)

\S4method{!=}{nanoperiod,nanoperiod}(e1, e2)
}
\arguments{
\item{x}{The object which want to convert to class \code{nanoperiod}}

\item{object}{argument for method \code{show}}

\item{...}{further arguments passed to or from methods.}

\item{i}{index specifying elements to extract or replace.}

\item{j}{Required for \code{[} signature but ignored here}

\item{drop}{Required for \code{[} signature but ignored here}

\item{e1}{Operand of class \code{nanoperiod}}

\item{e2}{Operand of class \code{nanoperiod}}
}
\description{
\code{nanoperiod} is a length of time type (implemented as an S4
class) with nanosecond precision. It differs from \code{nanoduration}
because it is capable of representing calendar months and days. It
can of course also reprensent years (immutably 12 months) and
weeks (immutably 7 days).
}
\section{Output Format}{


A \code{nanoperiod} is displayed as months, days, and \code{nanoduration}
like this: \code{10m2d/10:12:34.123_453_000}.
}

\section{Details}{




Adding or subtracting \code{nanoperiod} and \code{nanotime} require a
timezone as third argument. For this reason it is not possible to
use the binary operator `code{+}`. Instead the functions
`\code{plus}` and `\code{minus}` are defined.
}

\examples{
x <- nanoperiod(months=12, days=7, duration="01:00:00")
print(x)
y <- nanotime("1970-01-01T00:00:00.000000001+00:00")
plus(y, x, tz="America/Chicago")

}
\author{
Leonardo Silvestri
}
