% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\docType{class}
\name{period-class}
\alias{period-class}
\alias{period}
\alias{as.period,character-method}
\alias{as.period,integer64-method}
\alias{as.period,numeric-method}
\alias{as.period,integer-method}
\alias{as.period,duration-method}
\alias{show,period-method}
\alias{print,period-method}
\alias{as.character,period-method}
\alias{[,period,logical-method}
\alias{[,period,numeric-method}
\alias{[,period,character-method}
\alias{[<-,period,logical,ANY,period-method}
\alias{[<-,period,numeric,ANY,period-method}
\alias{[<-,period,character,ANY,period-method}
\alias{c.period}
\alias{names,period-method}
\alias{names<-,period-method}
\alias{Ops,period,ANY-method}
\alias{Ops,ANY,period-method}
\alias{Math,period-method}
\alias{Math2,period-method}
\alias{Summary,period-method}
\alias{Complex,period-method}
\alias{-,period,ANY-method}
\alias{-,period,period-method}
\alias{-,period,ANY-method}
\alias{-,period,duration-method}
\alias{-,period,integer64-method}
\alias{-,period,numeric-method}
\alias{-,ANY,period-method}
\alias{+,period,ANY-method}
\alias{+,period,period-method}
\alias{+,ANY,period-method}
\alias{+,period,duration-method}
\alias{+,period,integer64-method}
\alias{+,period,numeric-method}
\alias{+,period,ANY-method}
\alias{+,duration,period-method}
\alias{+,integer64,period-method}
\alias{+,numeric,period-method}
\alias{*,ANY,period-method}
\alias{*,period,ANY-method}
\alias{*,period,integer64-method}
\alias{*,period,numeric-method}
\alias{*,integer64,period-method}
\alias{*,numeric,period-method}
\alias{/,ANY,period-method}
\alias{/,period,ANY-method}
\alias{/,period,integer64-method}
\alias{/,period,numeric-method}
\alias{Compare,ANY,period-method}
\alias{Compare,period,ANY-method}
\alias{==,ANY,period-method}
\alias{==,period,ANY-method}
\alias{!=,ANY,period-method}
\alias{!=,period,ANY-method}
\alias{==,period,period-method}
\alias{!=,period,period-method}
\title{Period type with nanosecond precision}
\usage{
period(months, days, duration = as.duration(0))

\S4method{as.period}{character}(x)

\S4method{as.period}{integer64}(x)

\S4method{as.period}{numeric}(x)

\S4method{as.period}{integer}(x)

\S4method{as.period}{duration}(x)

\S4method{show}{period}(object)

\S4method{print}{period}(x, ...)

\S4method{as.character}{period}(x)

\S4method{[}{period,logical}(x, i, j, ..., drop = FALSE)

\S4method{[}{period,numeric}(x, i, j, ..., drop = FALSE)

\S4method{[}{period,character}(x, i, j, ..., drop = FALSE)

\S4method{[}{period,logical,ANY,period}(x, i, j, ...) <- value

\S4method{[}{period,numeric,ANY,period}(x, i, j, ...) <- value

\S4method{[}{period,character,ANY,period}(x, i, j, ...) <- value

\method{c}{period}(...)

\S4method{names}{period}(x)

\S4method{names}{period}(x) <- value

\S4method{Ops}{period,ANY}(e1, e2)

\S4method{Ops}{ANY,period}(e1, e2)

\S4method{Math}{period}(x)

\S4method{Math2}{period}(x, digits)

\S4method{Summary}{period}(x, ..., na.rm = FALSE)

\S4method{Complex}{period}(z)

\S4method{-}{period,ANY}(e1, e2)

\S4method{-}{period,period}(e1, e2)

\S4method{-}{period,ANY}(e1, e2)

\S4method{-}{period,duration}(e1, e2)

\S4method{-}{period,integer64}(e1, e2)

\S4method{-}{period,numeric}(e1, e2)

\S4method{-}{ANY,period}(e1, e2)

\S4method{+}{period,ANY}(e1, e2)

\S4method{+}{period,period}(e1, e2)

\S4method{+}{ANY,period}(e1, e2)

\S4method{+}{period,duration}(e1, e2)

\S4method{+}{period,integer64}(e1, e2)

\S4method{+}{period,numeric}(e1, e2)

\S4method{+}{period,ANY}(e1, e2)

\S4method{+}{duration,period}(e1, e2)

\S4method{+}{integer64,period}(e1, e2)

\S4method{+}{numeric,period}(e1, e2)

\S4method{*}{ANY,period}(e1, e2)

\S4method{*}{period,ANY}(e1, e2)

\S4method{*}{period,integer64}(e1, e2)

\S4method{*}{period,numeric}(e1, e2)

\S4method{*}{integer64,period}(e1, e2)

\S4method{*}{numeric,period}(e1, e2)

\S4method{/}{ANY,period}(e1, e2)

\S4method{/}{period,ANY}(e1, e2)

\S4method{/}{period,integer64}(e1, e2)

\S4method{/}{period,numeric}(e1, e2)

\S4method{Compare}{ANY,period}(e1, e2)

\S4method{Compare}{period,ANY}(e1, e2)

\S4method{==}{ANY,period}(e1, e2)

\S4method{==}{period,ANY}(e1, e2)

\S4method{!=}{ANY,period}(e1, e2)

\S4method{!=}{period,ANY}(e1, e2)

\S4method{==}{period,period}(e1, e2)

\S4method{!=}{period,period}(e1, e2)
}
\arguments{
\item{x}{something}
}
\description{
\code{period} is a length of time type (implemented as an S4
class) with nanosecond precision. It differs from \code{duration}
because it is capable of representing calendar months and days. It
can of course also reprensent years (immutably 12 months) and
weeks (immutably 7 days).
}
\section{Output Format}{


A \code{period} is displayed as months, days, and \code{duration}
like this: \code{10m2d/10:12:34.123_453_000}.
}

