% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoival.R
\name{rep,nanoival-method}
\alias{rep,nanoival-method}
\title{Replicate Elements}
\usage{
\S4method{rep}{nanoival}(x, ...)
}
\arguments{
\item{x}{a vector of \code{nanoival}}

\item{...}{further arguments:

       'times' an integer-valued vector giving the (non-negative)
           number of times to repeat each element if of length
           'length(x)', or to repeat the whole vector if of length
           1. Negative or 'NA' values are an error. A 'double'
           vector is accepted, other inputs being coerced to an
           integer or double vector.

       'length.out' non-negative integer. The desired length of the
           output vector. Other inputs will be coerced to a double
           vector and the first element taken. Ignored if 'NA' or
           invalid.

       'each' non-negative integer. Each element of 'x' is repeated
           'each' times.  Other inputs will be coerced to an integer
           or double vector and the first element taken. Treated as
           '1' if 'NA' or invalid.}
}
\description{
Replicates the values in 'x' similarly to the default method.
}
