% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanoival.R
\name{sort,nanoival-method}
\alias{sort,nanoival-method}
\title{Sorting or Ordering Vectors}
\usage{
\S4method{sort}{nanoival}(x, decreasing = FALSE)
}
\arguments{
\item{x}{a vector of \code{nanoival}}

\item{decreasing}{logical.  Should the sort be increasing or
decreasing?}
}
\description{
Sort (or _order_) a vector of \code{nanoival} into ascending or
descending order
}
\seealso{
\code{\link{is.unsorted}}
}
